better timeout functionality for parsl
